+++ lib/src/md/data/models/enums/enums.dart
@@ -1,1 +1,2 @@
 enum WantedAction { insert, update, delete }
+enum OtpType {email,sms}