+++ lib/src/md/data/models/files/upload_flie_model.dart
@@ -1,9 +1,9 @@
 import 'dart:convert';
 
 
-import '../../../../md_framework.dart';
-import '../../../api_constants.dart';
-import '../../encryption.dart';
+import '../../../../../md_framework.dart';
+import '../../../../api_constants.dart';
+import '../../../helper/encryption.dart';
 
 class UploadFileModel {
   late WantedAction wantedAction;