+++ lib/src/md/data/models/request/md_request_model.dart
@@ -1,16 +1,22 @@
-import '../../../../md_framework.dart';
-import '../../../api_constants.dart';
+import '../../../../../md_framework.dart';
+import '../../../../api_constants.dart';
 
 class MDRequest {
   late String data;
   late String file;
-  MDRequest({required this.data, this.file = ""});
+
+  MDRequest(
+      {required this.data,
+      this.file = "",
+
+      });
 
   toJson() {
     return {
       "ApiToken": MD<ApiConstants>().apiToken,
       "Data": data,
       "encode_plc1": file,
+
     };
   }
 }