+++ lib/src/md/data/models/response/md_response_model.g.dart
@@ -9,9 +9,11 @@ part of 'md_response_model.dart';
 MDResponse _$MDResponseFromJson(Map<String, dynamic> json) => MDResponse()
   ..status = json['Result'] as String
   ..message = json['Error'] as String
-  ..data = json['Data'] as String?
+  ..data = json['Data'] as dynamic
   ..fileID = json['FileId'] as String?
   ..fileName = json['SavedFileName'] as String?
   ..file64 = json['FileData'] as String?
+  ..otpToken= json['TransToken'] as String?
   ..fileExtension = json['FileExt'] as String?;
+