+++ lib/src/network/api_error_handler.dart
@@ -8,7 +8,7 @@ class ApiErrorHandler {
     } else if (d.response!.statusCode == 401) {
       return MDResponse.a(status: '401', message: 'Unauthorized');
     } else if (d.response!.statusCode == 500) {
-      return MDResponse.a(status: '500', message: 'Internal Server Error');
+      return MDResponse.a(status: '500', message: 'Connection Error');
     } else if (d.response!.statusCode == 404) {
       return MDResponse.a(status: '404', message: 'Not Found');
     } else {